move ConditionalVectorFieldEstimator to base
