+++ sbi/neural_nets/estimators/__init__.py
@@ -1,5 +1,6 @@
 from sbi.neural_nets.estimators.base import (
     ConditionalDensityEstimator,
+    ConditionalVectorFieldEstimator,
     UnconditionalDensityEstimator,
 )
 from sbi.neural_nets.estimators.categorical_net import (
@@ -10,7 +11,4 @@ from sbi.neural_nets.estimators.flowmatching_estimator import FlowMatchingEstima
 from sbi.neural_nets.estimators.mixed_density_estimator import MixedDensityEstimator
 from sbi.neural_nets.estimators.nflows_flow import NFlowsFlow
 from sbi.neural_nets.estimators.score_estimator import ConditionalScoreEstimator
-from sbi.neural_nets.estimators.vector_field_estimator import (
-    ConditionalVectorFieldEstimator,
-)
 from sbi.neural_nets.estimators.zuko_flow import ZukoFlow, ZukoUnconditionalFlow
